compose-labs/employee-management/requirements.txt
Flask==3.1.3
pymongo==4.16.0
python-dotenv==1.2.1
